uploader/requirements.txt
google-api-python-client==2.110.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
instagrapi>=2.0.0
tenacity==8.2.3
loguru==0.7.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.0
click==8.1.7